+++ src/services/movieService.js
@@ -63,43 +63,3 @@ export const getMovies = async (movieIds) => {
     throw error;
   }
 };
-
-export const getMovieCast = async (movieId) => {
-  try {
-    const response = await fetch(`${API_BASE_URL}/cast/${movieId}`);
-    const data = await response.json();
-    return data;
-  } catch (error) {
-    throw error;
-  }
-};
-
-export const getMovieDirectors = async (movieId) => {
-  try {
-    const response = await fetch(`${API_BASE_URL}/cast/${movieId}/directors`);
-    const data = await response.json();
-    return data;
-  } catch (error) {
-    throw error;
-  }
-};
-
-export const getActorDetails = async (actorId) => {
-  try {
-    const response = await fetch(`${API_BASE_URL}/cast/${actorId}/details`);
-    const data = await response.json();
-    return data;
-  } catch (error) {
-    throw error;
-  }
-};
-
-export const getActorMovies = async (actorId) => {
-  try {
-    const response = await fetch(`${API_BASE_URL}/cast/${actorId}/movies`);
-    const data = await response.json();
-    return data;
-  } catch (error) {
-    throw error;
-  }
-};
